Move some code from ConfirmPage
