+++ src/app/ConfirmPage.tsx
@@ -1,6 +1,5 @@
 import * as React from "react";
 import classNames from "clsx";
-import { ThanosDAppMetadata } from "@thanos-wallet/dapp/dist/types";
 import { useLocation } from "lib/woozie";
 import {
   useThanosClient,
@@ -21,7 +20,6 @@ import AccountBanner from "app/templates/AccountBanner";
 import NetworkBanner from "app/templates/NetworkBanner";
 import Balance from "app/templates/Balance";
 import CustomSelect, { OptionRenderProps } from "app/templates/CustomSelect";
-import Logo from "app/atoms/Logo";
 import Identicon from "app/atoms/Identicon";
 import Name from "app/atoms/Name";
 import AccountTypeBadge from "app/atoms/AccountTypeBadge";
@@ -31,11 +29,10 @@ import FormSubmitButton from "app/atoms/FormSubmitButton";
 import FormSecondaryButton from "app/atoms/FormSecondaryButton";
 import ConfirmLedgerOverlay from "app/atoms/ConfirmLedgerOverlay";
 import HashShortView from "app/atoms/HashShortView";
-import { ReactComponent as ComponentIcon } from "app/icons/component.svg";
-import { ReactComponent as OkIcon } from "app/icons/ok.svg";
-import { ReactComponent as LayersIcon } from "app/icons/layers.svg";
 import DAppLogo from "app/templates/DAppLogo";
 import OperationView from "app/ConfirmPage/OperationView";
+import SubTitle from "app/ConfirmPage/SubTitle";
+import ConnectBanner from "app/ConfirmPage/ConnectBanner";
 
 const ConfirmPage: React.FC = () => {
   const { ready } = useThanosClient();
@@ -475,105 +472,4 @@ const AccountOptionContentHOC = (networkRpc: string) => {
       </div>
     </>
   ));
-};
-
-type ConnectBannerProps = {
-  type: "connect" | "confirm_operations";
-  origin: string;
-  appMeta: ThanosDAppMetadata;
-  className?: string;
-};
-
-const ConnectBanner: React.FC<ConnectBannerProps> = ({
-  type,
-  origin,
-  appMeta,
-  className,
-}) => {
-  const Icon = type === "connect" ? OkIcon : LayersIcon;
-
-  return (
-    <div
-      className={classNames(
-        "w-full flex items-center justify-around",
-        className
-      )}
-    >
-      <div
-        className={classNames(
-          "w-32",
-          "border border-gray-200 rounded",
-          "flex flex-col items-center",
-          "p-2"
-        )}
-      >
-        <DAppLogo origin={origin} size={32} className="flex-shrink-0 mb-1" />
-
-        <span className="text-xs font-semibold text-gray-700">
-          <Name style={{ maxWidth: "7.5rem" }}>{appMeta.name}</Name>
-        </span>
-      </div>
-
-      <div className="relative flex-1 h-px bg-gray-300">
-        <div className="absolute inset-0 flex items-center justify-center">
-          <div
-            className={classNames(
-              type === "connect" ? "bg-green-500" : "bg-orange-500",
-              "rounded-full",
-              "p-1",
-              "flex items-center justify-center",
-              "text-white"
-            )}
-          >
-            <Icon className="w-auto h-4 stroke-current stroke-2" />
-          </div>
-        </div>
-      </div>
-
-      <div
-        className={classNames(
-          "w-32",
-          "border border-gray-200 rounded",
-          "flex flex-col items-center",
-          "p-2"
-        )}
-      >
-        <Logo className="mb-1" imgStyle={{ height: 32, margin: "auto" }} />
-
-        <span className="text-xs font-semibold text-gray-700">Thanos</span>
-      </div>
-    </div>
-  );
-};
-
-type SubTitleProps = React.HTMLAttributes<HTMLHeadingElement>;
-
-const SubTitle: React.FC<SubTitleProps> = ({
-  className,
-  children,
-  ...rest
-}) => {
-  const comp = (
-    <span className="px-1 text-gray-500">
-      <ComponentIcon className="w-auto h-5 stroke-current" />
-    </span>
-  );
-
-  return (
-    <h2
-      className={classNames(
-        "flex items-center justify-center",
-        "text-gray-700",
-        "text-lg",
-        "font-light",
-        "uppercase",
-        className
-      )}
-      {...rest}
-    >
-      {comp}
-      {children}
-      {comp}
-    </h2>
-  );
 };
