+++ src/app/ConfirmPage/SubTitle.tsx
@@ -1,0 +1,37 @@
+import classNames from "clsx";
+import React from "react";
+import { ReactComponent as ComponentIcon } from "app/icons/component.svg";
+
+type SubTitleProps = React.HTMLAttributes<HTMLHeadingElement>;
+
+const SubTitle: React.FC<SubTitleProps> = ({
+  className,
+  children,
+  ...rest
+}) => {
+  const comp = (
+    <span className="px-1 text-gray-500">
+      <ComponentIcon className="w-auto h-5 stroke-current" />
+    </span>
+  );
+
+  return (
+    <h2
+      className={classNames(
+        "flex items-center justify-center",
+        "text-gray-700",
+        "text-lg",
+        "font-light",
+        "uppercase",
+        className
+      )}
+      {...rest}
+    >
+      {comp}
+      {children}
+      {comp}
+    </h2>
+  );
+};
+
+export default SubTitle;
